Update README.md
This repository contains a convolutional autoencoder built with TensorFlow and Keras for unsupervised learning on the MNIST dataset. The autoencoder learns to compress and reconstruct grayscale images of handwritten digits. It consists of convolutional layers in both the encoder and decoder parts, optimized using binary crossentropy loss. The training set is split into training and validation subsets to monitor overfitting, while the test set is used only for final evaluation.